# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: NebulaCloud API
  version: "0.0.1"
  description: |
    NebulaCloud is a self-hosted PaaS in the spirit of Railway, Render, Heroku
    and Vercel. This document describes the public REST + WebSocket surface
    consumed by the dashboard and by automation clients.

    Phase 0 shipped health and meta. Phase 1 adds Identity. Phase 2 adds
    organizations, projects, services, environment variables, deployments,
    GitHub webhooks, and scaffolded logs/metrics queries for the dashboard.

    Subsequent phases will deepen Domains, streaming logs, and richer metrics.

  contact:
    name: NebulaCloud
    url: https://github.com/nebulacloud/nebula
servers:
  - url: http://api.nebula.localhost
    description: local development
  - url: https://api.example.com
    description: production (replace with your domain)
...
paths:
...
  /api/v1/webhooks/github:
    post:
      tags: [Webhooks]
      operationId: githubWebhook
      summary: GitHub push webhook (HMAC optional in dev)
      description: |
        Accepts GitHub `push` and `ping` events. When `X-GitHub-Event` is
        something else, responds with `{data:{status:"ignored"}}`. Signature
        enforced when `X-Hub-Signature-256` matches the configured secret.
        Push payloads may include `installation.id`; when present and the
        project has `github_installation_id` set, only matching projects are
        deployed (otherwise URL + default-branch matching applies as in dev).
      parameters:
        - name: X-GitHub-Event
          in: header
          required: false
          schema:
            type: string
        - name: X-Hub-Signature-256
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Processed or ignored event.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    additionalProperties: true
        "204":
          description: Ping acknowledged.
        "401":
          description: Bad or missing webhook signature when secret is enforced.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Invalid JSON or payload.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/services/{serviceID}/domains:
    get:
      tags: [Domains]
      operationId: listDomains
      summary: List custom domains for a service
      security:
        - BearerAuth: []
      parameters:
        - name: serviceID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Domain rows.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Domain" }
    post:
      tags: [Domains]
      operationId: createDomain
      summary: Register a custom hostname (pending verification)
      security:
        - BearerAuth: []
      parameters:
        - name: serviceID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostname]
              properties:
                hostname: { type: string }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Domain" }
  /api/v1/services/{serviceID}/domains/{domainID}:
    delete:
      tags: [Domains]
      operationId: deleteDomain
      summary: Remove a custom domain
      security:
        - BearerAuth: []
      parameters:
        - name: serviceID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: domainID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted.
  /api/v1/services/{serviceID}/domains/{domainID}/verify:
    post:
      tags: [Domains]
      operationId: verifyDomain
      summary: Mark domain verified (ACME on next deploy)
      security:
        - BearerAuth: []
      parameters:
        - name: serviceID
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: domainID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Verified domain.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: "#/components/schemas/Domain" }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Domain:
      type: object
      required: [id, service_id, hostname, ssl_status, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        service_id: { type: string, format: uuid }
        hostname: { type: string }
        is_primary: { type: boolean }
        ssl_status: { type: string, enum: [pending, issued, failed] }
        verification_token: { type: string }
        verified_at: { type: string, format: date-time }
        cname_hint: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [kind, message]
          properties:
            kind:
              type: string
              enum:
                - validation
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - rate_limited
                - unavailable
                - internal
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: { type: string }